# openapi.json
{
  "components": {
    "schemas": {
...
      "AddMovieIn": {
        "properties": {
          "monitored": {
            "default": true,
            "title": "Monitored",
            "type": "boolean"
          },
          "quality_profile_id": {
            "title": "Quality Profile Id",
            "type": "integer"
          },
          "root_folder_path": {
            "title": "Root Folder Path",
            "type": "string"
          },
          "search_now": {
            "default": true,
            "title": "Search Now",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tmdb_id": {
            "title": "Tmdb Id",
            "type": "integer"
          }
        },
        "required": [
          "tmdb_id",
          "title",
          "quality_profile_id",
          "root_folder_path"
        ],
        "title": "AddMovieIn",
        "type": "object"
      },
      "AddSeriesIn": {
        "properties": {
          "monitored": {
            "default": true,
            "title": "Monitored",
            "type": "boolean"
          },
          "quality_profile_id": {
            "title": "Quality Profile Id",
            "type": "integer"
          },
          "root_folder_path": {
            "title": "Root Folder Path",
            "type": "string"
          },
          "search_now": {
            "default": true,
            "title": "Search Now",
            "type": "boolean"
          },
          "season_folder": {
            "default": true,
            "title": "Season Folder",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tvdb_id": {
            "title": "Tvdb Id",
            "type": "integer"
          }
        },
        "required": [
          "tvdb_id",
          "title",
          "quality_profile_id",
          "root_folder_path"
        ],
        "title": "AddSeriesIn",
        "type": "object"
      },
...
      "CollectionDetailOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "monitored": {
            "default": false,
            "title": "Monitored",
            "type": "boolean"
          },
          "movies": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SearchResultOut"
            },
            "title": "Movies",
            "type": "array"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "poster": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id"
        ],
        "title": "CollectionDetailOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptionsOut": {
        "properties": {
          "metadata_profiles": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/QualityProfileOut"
            },
            "title": "Metadata Profiles",
            "type": "array"
          },
          "quality_profiles": {
            "items": {
              "$ref": "#/components/schemas/QualityProfileOut"
            },
            "title": "Quality Profiles",
            "type": "array"
          },
          "root_folders": {
            "items": {
              "$ref": "#/components/schemas/RootFolderOut"
            },
            "title": "Root Folders",
            "type": "array"
          }
        },
        "required": [
          "quality_profiles",
          "root_folders"
        ],
        "title": "OptionsOut",
        "type": "object"
      },
...
      "QualityProfileOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "QualityProfileOut",
        "type": "object"
      },
...
      "RootFolderOut": {
        "properties": {
          "free_space": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Space"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "id",
          "path"
        ],
        "title": "RootFolderOut",
        "type": "object"
      },
...
      "SearchResultOut": {
        "properties": {
          "has_file": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has File"
          },
          "imdb_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imdb Id"
          },
          "in_library": {
            "default": false,
            "title": "In Library",
            "type": "boolean"
          },
          "kind": {
            "enum": [
              "movie",
              "series"
            ],
            "title": "Kind",
            "type": "string"
          },
          "library_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library Id"
          },
          "monitored": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monitored"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "poster": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster"
          },
          "quality_profile_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Profile Id"
          },
          "remote_id": {
            "title": "Remote Id",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tmdb_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb Id"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "kind",
          "title",
          "remote_id"
        ],
        "title": "SearchResultOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "arrdeck",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/collections/{collection_id}": {
      "get": {
        "operationId": "collection_detail_api_v1_collections__collection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "title": "Collection Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Collection Detail",
        "tags": [
          "discover"
        ]
      },
      "patch": {
        "operationId": "toggle_collection_api_v1_collections__collection_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "title": "Collection Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "monitored",
            "required": true,
            "schema": {
              "title": "Monitored",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Toggle Collection Api V1 Collections  Collection Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Toggle Collection",
        "tags": [
          "discover"
        ]
      }
    },
...
    "/api/v1/discover/movies": {
      "get": {
        "operationId": "discover_movies_api_v1_discover_movies_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResultOut"
                  },
                  "title": "Response Discover Movies Api V1 Discover Movies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover Movies",
        "tags": [
          "discover"
        ]
      }
    },
    "/api/v1/discover/series": {
      "get": {
        "operationId": "discover_series_api_v1_discover_series_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResultOut"
                  },
                  "title": "Response Discover Series Api V1 Discover Series Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover Series",
        "tags": [
          "discover"
        ]
      }
    },
...
    "/api/v1/movies": {
      "post": {
        "operationId": "add_movie_api_v1_movies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMovieIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Movie Api V1 Movies Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Movie",
        "tags": [
          "discover"
        ]
      }
    },
    "/api/v1/options/{app}": {
      "get": {
        "operationId": "options_api_v1_options__app__get",
        "parameters": [
          {
            "in": "path",
            "name": "app",
            "required": true,
            "schema": {
              "title": "App",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Options",
        "tags": [
          "discover"
        ]
      }
    },
...
    "/api/v1/search/movies": {
      "get": {
        "operationId": "search_movies_api_v1_search_movies_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResultOut"
                  },
                  "title": "Response Search Movies Api V1 Search Movies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Movies",
        "tags": [
          "discover"
        ]
      }
    },
...
    "/api/v1/search/series": {
      "get": {
        "operationId": "search_series_api_v1_search_series_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResultOut"
                  },
                  "title": "Response Search Series Api V1 Search Series Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Series",
        "tags": [
          "discover"
        ]
      }
    },
    "/api/v1/series": {
      "post": {
        "operationId": "add_series_api_v1_series_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSeriesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Series Api V1 Series Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Series",
        "tags": [
          "discover"
        ]
      }
    },
...
  }
}